Refactor: split out IntroScreen component
Closes: #28

- Moves all concerns about the top nav bar (first screenshot below) away from the intro screen component (second screenshot below).  New component is called 'TopNavBar'

- added component tests to IntroScreen

- fixed the barrel file for RoomsSquare
