+++ src/client/components/NameChanger/NameChanger.tsx
@@ -7,51 +7,6 @@ interface NameChangerProps {
     handleSubmit: (name: string) => void;
 }
 
-const NameChangerContainer = styled.div`
-    width: 700px;
-    margin: auto;
-    height: 100vh;
-    display: grid;
-    place-items: center;
-    animation: fadein 1s;
-
-    @keyframes fadein {
-        from {
-            opacity: 0.01;
-        }
-        to {
-            opacity: 1;
-        }
-    }
-    @-moz-keyframes fadein {
-        /* Firefox */
-        from {
-            opacity: 0.01;
-        }
-        to {
-            opacity: 1;
-        }
-    }
-    @-webkit-keyframes fadein {
-        /* Safari and Chrome */
-        from {
-            opacity: 0.01;
-        }
-        to {
-            opacity: 1;
-        }
-    }
-    @-o-keyframes fadein {
-        /* Opera */
-        from {
-            opacity: 0.01;
-        }
-        to {
-            opacity: 1;
-        }
-    }
-`;
-
 const NameChangerForm = styled.form`
     display: grid;
     grid-gap: 12px;
@@ -84,33 +39,31 @@ export const NameChanger: React.FC<NameChangerProps> = ({ handleSubmit }) => {
         handleSubmit(name);
     };
     return (
-        <NameChangerContainer>
-            <NameChangerForm onSubmit={onSubmit}>
-                <div>
-                    <label htmlFor="name-selector">Choose a Name</label>
-                    <input
-                        id="name-selector"
-                        role="textbox"
-                        maxLength={MAX_PLAYER_NAME_LENGTH}
-                        autoFocus
-                        autoComplete="off" // ignore for password managers
-                        data-lpignore="true" // ignore for lastPass
-                        data-form-type="other" // ignore for dashlane
-                        value={name}
-                        onChange={(e) => setName(e.target.value)}
-                    />
-                </div>
-                <div>
-                    <PrimaryColorButton
-                        emoji="▶️"
-                        role="button"
-                        type="submit"
-                        disabled={!name}
-                    >
-                        Start
-                    </PrimaryColorButton>
-                </div>
-            </NameChangerForm>
-        </NameChangerContainer>
+        <NameChangerForm onSubmit={onSubmit}>
+            <div>
+                <label htmlFor="name-selector">Choose a Name</label>
+                <input
+                    id="name-selector"
+                    role="textbox"
+                    maxLength={MAX_PLAYER_NAME_LENGTH}
+                    autoFocus
+                    autoComplete="off" // ignore for password managers
+                    data-lpignore="true" // ignore for lastPass
+                    data-form-type="other" // ignore for dashlane
+                    value={name}
+                    onChange={(e) => setName(e.target.value)}
+                />
+            </div>
+            <div>
+                <PrimaryColorButton
+                    emoji="▶️"
+                    role="button"
+                    type="submit"
+                    disabled={!name}
+                >
+                    Start
+                </PrimaryColorButton>
+            </div>
+        </NameChangerForm>
     );
 };
