+++ src/client/components/RoomSquare/index.ts
@@ -1,0 +1,1 @@
+export * from './RoomSquare';
