+++ src/client/components/TopNavBar/TopNavBar.spec.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import { fireEvent, screen } from '@testing-library/react';
+
+import { render } from '@/test-utils';
+import { TopNavBar } from './TopNavBar';
+import { RootState } from '@/client/redux/store';
+
+describe('TopNavBar', () => {
+    it('logs out', () => {
+        const preloadedState: Partial<RootState> = {
+            user: { name: 'Squidcake' },
+        };
+        const { webSocket } = render(<TopNavBar />, { preloadedState });
+
+        fireEvent.click(screen.getByText('Logout'));
+
+        expect(webSocket.chooseName).toHaveBeenCalledWith('');
+    });
+});
