+++ src/client/components/TopNavBar/index.ts
@@ -1,0 +1,1 @@
+export * from './TopNavBar';
